pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "py-conformational-sampling"
version = "0.6.0"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    { name = "Joshua Kammeraad" },
    { name = "Soumik Das"},
]
dependencies = [
    "ase",
    "hvplot",
    "morfeus-ml",
    "nglview",
    "numpy<2",
    "openbabel-wheel",
    "pandas",
    "panel-chemistry",
    "panel<1.0",
    "param",
    "pyGSM @ git+https://github.com/ZimmermanGroup/pyGSM.git",
    "pytest",
    "rdkit",
    "setuptools", # necessary for pyGSM to run on python 3.11
    "stk<=2021.8.2.0",
    "stko<=0.0.40",
    "xtb",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "dask-jobqueue",
    "snakeviz",
]
vscode = [
    "jupyter",
    "pylint",
]

[project.urls]
Homepage = "https://github.com/ZimmermanGroup/py-conformational-sampling"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff.format]
quote-style = "single"
